requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
google-api-python-client==2.129.0
google-auth==2.32.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
